third_party/blink/web_tests/fast/events/event-properties-gc-expected.txt

Tests that properties passed to events are not garbage collected prematurely.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS event[prop] is "foo"
PASS event[prop] is "foo"
PASS event[prop] is "foo"
PASS successfullyParsed is true

TEST COMPLETE
